Task 00 - Armazem de itens

+++ 0x22/armazem/Comida.java
@@ -1,7 +1,7 @@
 public class Comida {
 
     private String nome;
-    private int calorias;
+    private double calorias;
     private double preco;
 
     public Comida(String nome, int calorias, double preco) {
@@ -18,11 +18,11 @@ public class Comida {
         this.nome = nome;
     }
 
-    public int getCalorias() {
+    public double getCalorias() {
         return calorias;
     }
 
-    public void setCalorias(int calorias) {
+    public void setCalorias(double calorias) {
         this.calorias = calorias;
     }
 
@@ -36,7 +36,7 @@ public class Comida {
 
     @Override
     public String toString() {
-        return String.format("[%s] %d R$ %f",
+        return String.format("[%s] %f R$ %f",
                 getNome(), getCalorias(), getPreco());
     }
 }
